refactor: simplify OffSiteGatewayWebhookNotificationHandler class
